+++ src/Compiler/Services/BaseClassLibraryReferences.cs
@@ -120,13 +120,12 @@ namespace Jroc.Services
         public TypeReferenceHandle Int32Type => _typeRefRegistry.GetOrAdd(typeof(int));
         public TypeReferenceHandle ObjectType => _typeRefRegistry.GetOrAdd(typeof(object));
         public TypeReferenceHandle StringType => _typeRefRegistry.GetOrAdd(typeof(string));
-        public TypeReferenceHandle ExceptionType => _typeRefRegistry.GetOrAdd(typeof(System.Exception));
         public TypeReferenceHandle SystemMathType => _typeRefRegistry.GetOrAdd(typeof(System.Math));
         public TypeReferenceHandle MethodBaseType => _typeRefRegistry.GetOrAdd(typeof(System.Reflection.MethodBase));
         public TypeReferenceHandle JsFunctionObjectType => _typeRefRegistry.GetOrAdd(typeof(JavaScriptRuntime.JsFunctionObject));
-        public TypeReferenceHandle JsThrownValueExceptionType =>
+        public TypeReferenceHandle JsAsyncFunctionObjectType =>
             _typeRefRegistry.GetOrAdd(
-                typeof(JavaScriptRuntime.JsThrownValueException));
+                typeof(JavaScriptRuntime.JsAsyncFunctionObject));
         public TypeReferenceHandle JsCallArgumentsType => _typeRefRegistry.GetOrAdd(typeof(JavaScriptRuntime.JsCallArguments));
         public TypeReferenceHandle InAttributeType => _typeRefRegistry.GetOrAdd(typeof(System.Runtime.InteropServices.InAttribute));
         public EntityHandle ObjectArrayType => _memberRefRegistry.GetOrAddTypeHandle(typeof(object[]));
@@ -134,6 +133,10 @@ namespace Jroc.Services
         public MemberReferenceHandle Object_Ctor_Ref => _memberRefRegistry.GetOrAddConstructor(typeof(object), Type.EmptyTypes);
         public MemberReferenceHandle JsObject_Ctor_Ref => _memberRefRegistry.GetOrAddConstructor(typeof(JavaScriptRuntime.JsObject), Type.EmptyTypes);
         public MemberReferenceHandle JsFunctionObject_Ctor_Ref => _memberRefRegistry.GetOrAddConstructor(typeof(JavaScriptRuntime.JsFunctionObject), Type.EmptyTypes);
+        public MemberReferenceHandle JsAsyncFunctionObject_Ctor_Ref =>
+            _memberRefRegistry.GetOrAddConstructor(
+                typeof(JavaScriptRuntime.JsAsyncFunctionObject),
+                Type.EmptyTypes);
         public MemberReferenceHandle JsCallArguments_GetArgument_Ref => _memberRefRegistry.GetOrAddMethod(
             typeof(JavaScriptRuntime.JsCallArguments),
             nameof(JavaScriptRuntime.JsCallArguments.GetArgument),
@@ -167,11 +170,6 @@ namespace Jroc.Services
         public MemberReferenceHandle TypeError_Ctor_String_Ref => _memberRefRegistry.GetOrAddConstructor(
             typeof(JavaScriptRuntime.TypeError),
             new[] { typeof(string) });
-        public MemberReferenceHandle Promise_Reject_Object_Ref =>
-            _memberRefRegistry.GetOrAddMethod(
-                typeof(JavaScriptRuntime.Promise),
-                nameof(JavaScriptRuntime.Promise.reject),
-                new[] { typeof(object) });
         public MemberReferenceHandle GeneratorObject_InitializeGeneratorFunctionSurface_Ref =>
             _memberRefRegistry.GetOrAddMethod(
                 typeof(JavaScriptRuntime.GeneratorObject),
@@ -187,10 +185,6 @@ namespace Jroc.Services
                 typeof(JavaScriptRuntime.GeneratorObject),
                 nameof(JavaScriptRuntime.GeneratorObject.InitializeInstanceFromFunction),
                 new[] { typeof(object), typeof(object) });
-        public MemberReferenceHandle JsThrownValueException_Value_Getter_Ref =>
-            _memberRefRegistry.GetOrAddMethod(
-                typeof(JavaScriptRuntime.JsThrownValueException),
-                $"get_{nameof(JavaScriptRuntime.JsThrownValueException.Value)}");
         public MemberReferenceHandle RuntimeServices_ResolveLexicalThis_Ref =>
             _memberRefRegistry.GetOrAddMethod(
                 typeof(JavaScriptRuntime.RuntimeServices),